Adding SVC API and Corresponding helper
Added helper store_metadata_url for use in installer templates and
added the corresponding route.

Special note, for any helper/svc route, if the URL generated includes
any &'s, the a curl command in an installer must have the whole url
quoted.  Also using []'s in the uri (for submitting multiple params of
the same name and then being able to retrieve the array), the []'s need
to be escaped.