Add `check-migrations` command
This adds a command to the admin tool that will check if the database requires
a migration, or if it is fully up to date.  This helps support the Puppet
module robustly detecting if migrations need to be run -- and, especially, if
we need to restart the application to reflect those changes.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>