Change how create-policy maintains order of policies
The previous scheme of explicitly exposing users to the rule_number is very
cumbersome. It's more intuitive to allow users to not specify an order and
default to appending new policies, and if they do want to control order, do
that by expressing before/after which policy the new policy should go.

Fixes https://github.com/puppetlabs/razor-server/issues/137