Don't ask bundler to retry downloads
Turns out we don't have broad access to the `--retry` flag to bundler, which
is only available in very recent versions.  Theoretically we don't need it
anyway, so removing it is the short path.

Since that exposes us to some hassles caused by unreliable mirrors upstream,
we manually implement the retrying, and hope that fixes things.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>