Don't require JRuby to load the Rakefile
We depend on JRuby for one small part of the Rakefile, creating the vendor gem
bundle prior to packaging.  For the rest of the code we don't actually need
it, and raising an error if loaded with the wrong engine is
actually unhelpful.

Now we just let the commands fail later if appropriate.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>