Update os_boot.erb
Using the underscore character "_" for the separator can be problematic - if the <%= repo_url %> value contains an underscore .

My example, the repo was named "ubuntu_server" and of course the sed command would work out when hitting the underscore in the url (http://192.168.1.5:8080/svc/repo/ubuntu_server)

I replaced the delimiter with the character that has the octal value 001 - in ASCII it's the SOH character (start of heading) - which is extremely unlikely to appear in the repo name.

I think it is safer to use this.