Fix intermittent test failure due to content ordering
One of the IPMI tests would very occasionally fail because the order of data
was not what was expected.  This was a situation I considered theoretical,
rather than practical, when the test was first written -- and the order does
not matter, just the unordered set of values.

This updates the test to use an order-insensitive comparison, ensuring it will
always pass regardless of random chance altering the order.

https://tickets.puppetlabs.com/browse/RAZOR-177

Signed-off-by: Daniel Pittman <daniel@rimspace.net>