Validate `create-broker`, and make `references` mode liberal
This adds validation for the `create-broker` command.  Unlike previous
commands, this one has a field that references an object that is not stored in
the database, the BrokerType.

Supporting that made for two new features required in the validation code:
 * use a duck-type interface to check if the object exists
 * support different "search key" and "attribute name" in references

Finally, since we needed a unified interface for everything that we reference,
this updates the BrokerType `find` method to follow the same conventions as
the Sequel::Model classes.  That eliminates various special case code, and
lets us treat them polymorphicly.

https://tickets.puppetlabs.com/browse/RAZOR-118

Signed-off-by: Daniel Pittman <daniel@rimspace.net>