Update command status for background tasks
For ISO import, we thread a command object through to the message
processor, tkaing care to pass it along to each stage of the import
process. The import process then updates the status of the command, in
particular marks it as 'running' as soon as it starts working on it, and
marks it as 'finished' when the last step of the import saga is done.

Most of the error reporting is done by our message processor - that is
crude, and better ways to report errors from within the actual processing
is the main reason why we try to pass the command to each stage of the
saga.

We can't demand that all background tasks have a command associated with
it: changing node metadata and tag rules spawn background jobs in
Node.around_save and Tag.around_save, where they do not have access to a
command object, nor does it seem desirable from the user's POV to treat
them as async commands. There may be a case to give the user an indication that the database has settled after these changes, but that is unlikely to change the general situation that we'll have background work that is, to the user, part of their command, and some that isn't.

The way the command is threaded through all over the place is a bit clumsy;
an alternative would be to have the messaging plugin define a 'command'
method that returns the command associated with the operation if there is
one, and methods that are called during async processing can access it that
way to e.g. report errors or update the status