Validate `create-policy`, and add array attribute validation
This is a large change, because array attribute validation behaves somewhat
differently to hash attribute validation, resulting in the final split between
the two.

This was necessary to implement validation of the `create-policy` command,
which uses an array of nested objects.  Now that is done, we should have all
the major features required for validation in place.

https://tickets.puppetlabs.com/browse/RAZOR-118

Signed-off-by: Daniel Pittman <daniel@rimspace.net>