Implement a 'size' validation check for attributes
This implement a check to validate the size of an attribute; it only makes
sense for string, object/map, and array types.  It uses a range to verify the
minimum and maximum size of the object in characters, or contained entities.

This, in turn, supports verification that inappropriate length entities are
not passed in our data.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>