Add `register-node` command to pre-register nodes with Razor
In order to make brownfield deployments of Razor easier we allow users to
register nodes explicitly.  This command allows you to perform the same
registration that would happen when a new node checked in, ahead of time.

In order for this to be effective the hw_info must contain enough information
that the node can successfully be matched during the iPXE boot phase.

If the node matches an existing node, in keeping with the overall policy
of commands declaring desired state, the node installed field will be
updated to match the value in this command.

The final state is that a node with the supplied hardware information,
and the desired installed state, will be present in the database,
regardless of it existing before hand or not.

https://tickets.puppetlabs.com/browse/RAZOR-151

Signed-off-by: Daniel Pittman <daniel@rimspace.net>