ASM-2130 do not modify hw_info in /svc/nodeid
Windows install uses the /svc/nodeid REST endpoint to look up its node
id based on its mac addresses. However, the current implementation of
that service was modifying the node hw_info hash to match those passed
in the call. This was erasing the serial field that asm-deployer uses
to find nodes causing deployments to fail.

This issue seems to be mostly fixed upstream in razor 0.15.0. That
implementation still has a minor issue in that /svc/nodeid will create
a new node and return it if the passed parameters are not found, but
at least it doesn't clobber existing hw_info data.