ASM-3312 support fqdn hostnames
The razor task scripts were calculating the correct puppet agent
certificate name to use based off of the first word in the
hostname. This certificate name has to match that used in
asm-deployer in order for the correct puppet configuration to be
applied to the node. In the case where the FQDN was used as the
hostname the certificate name calculated on the razor side was
incorrect because it missed the domain name. It's not possible to
just use the fqdn when calculating the agent certificate name
because the original hostname specified by the user may not have
been an FQDN.

This commit adds agent_certname to the razor policy
installer_options. The razor task scripts use that option which
guarantees that they are using the same certificate name that
asm-deployer uses.