better (ASM-specific) repo task migration
The default razor migration scripts for setting the new repo
task_name field were not very good, at least for ASM. They only
set a repo task name if that repo had been used with exactly one
policy. In the ASM case we create a new policy for every server
so there is almost always more than one policy associated with a
repo if it was ever used. Additionally in the case of Windows,
the policy task name was a version-specific task, but ASM wants
the base windows task associated with the repo.

Note that in pathological cases where the same repo had been used
with disparate tasks (e.g. a repo used for both redhat and
windows2012 deployments) it will be indeterminate which task will
be set on the repo with this code. One of the tasks actually used
will win and be set on the repo.

- all repos that have been used with any policy using a task name
  starting with windows2012 or windows2008 will be migrated to
  windows2012 or windows2008 tasks. This is necessary because ASM
  includes the specific Windows version,
  e.g. windows2012r2datacenter, as the task associated with the
  policy rather than the expected base task name of just
  e.g. windows2012

- all repos that have been used with vmware_esxi, redhat or
  redhat7 tasks are migrated to those tasks.

- esxi-5.1 and esxi-5.5 repos are migrated to vmware_esxi task

- any remaining repos left with noop task are logged