ASM-5492 skip libselinux-ruby if already installed
Puppet installation on RHEL / CentOS 7.2 was failing because
libselinux-ruby was already installed. This commit checks if it is
already installed and skips it if so.

In future we need to move to using yum to install the puppet agent so
that problems like this don't occur. It would also allow us to support
minimal ISOs that don't include samba support.