ASM-6906 Do not set hosts entry if already set
The final step of the redhat post_install script is to remove itself
from /etc/rc.d/rc.local so that it doesn't get re-run when the server
next reboots. However it's possible for it to execute twice if the
server gets rebooted while it is running. In that case it was adding a
duplicate dellasm hosts file entry. That will cause problems later on
if e.g. you use puppet to manage that entry. In that case puppet will
only update the second of the entries and the server will continue to
use the first IP.