ASM-7185 Win2008R2 Bare-Metal Install Failing with Error "O/S install timed out"
Fixed issue where we were getting "Disk is not initialized error.  Put a
try catch around the code to format disk and lay down OS, so if a failure
occurs we don't notify razor that the OS has been installed.  The server
will then reboot and try again.  After the reboot the install seems to go
through OK.