ASM-7709 setup ntp before puppet config
This commit addresses the issue where the ntp service interferes
with the puppet-agent process on the Windows 2008 and 2012. When
the interference occurs, the puppet-agent is stuck indefinitely
until the system time catches up to new time set by the ntp service.
This typically takes a few hours to sync, so the deployment times
out and fails.

In order to resolve this issue, default-unattended.xml.erb files for
the Windows 2008 and 2012 are updated. The RunSynchronous commands
are rearranged to configure the ntp service before installing and
configuring the puppet service. Additionally, the ntp service is
set to start up at the system startup while the puppet service is
set to start after other system services (including the ntp) to
reduce the possibility of the interference.

The validation results show that this removed the deployment fail-
ures from the environment where ntp/puppet conflicts could be seen
for every deployment. Also deploying the Windows VM's without the
ntp server(s) also function correctly.