ASM- 8293 ESXi VSAN SATADOM deployment is installing ESXi on different disk
On VMware shell, script is executed as normal shell instead of bash shell. Script changes added recently was not working as a result the last ATA disk in the enumeration is selected for the OS installation. Updated the script based on "SHELL" to ensure we select the correct disk device.