ASM-8322 Allow VSAN deployment to RAID volume
The ESXi installer tries to determine the disk to use for OS
installation for VSAN_AHCI deployments by looking for disks with
names like "Local ATA Disk" or "DELL Serial". But PERC RAID volumes
are named "Local DELL Disk". Therefore the installer was failing to
find the install disk for VSAN_AHCI deployments to servers with RAID
volumes.