ASM-9173 Set Windows w32time start trigger
By default the Windows w32time (NTP) service if set to Auto start is
actually set to Auto (triggered). And the default trigger is when the
domain service starts. Therefore in cases where the server is not
joined to a domain the service never starts up after reboot. This
behavior and workaround is document in the Microsoft support article
[Windows Time service doesn't start automatically on a workgroup
computer that's running Windows 7 or Windows Server 2008 R2][1].

Additionally the timezone setting in the Windows unattended.xml file
seems to be applied after the `<RunSynchronous>` commands are
executed. We use the `<RunSycnhronous>` commands to configure
NTP. Therefore NTP starts (and presumably sets the correct time) and
then the timezone is changed which changes the system time. Since most
of the unattended.xml files set Central timezone and the default
appears to be Pacific, this results in the server time being two hours
behind where it should be. After restart of the server NTP is not
started (if the domain service is not configured) resulting in an
incorrect time forever after.

This change follows one of the recommendations in the linked article
to change the trigger to be when the network starts or stops rather
than when the domain service starts or stops. Therefore NTP should
start after reboot and correct the time.

Additionally it appears NTP settings were left out entirely of the
Windows 2008 Hyper-V unattend.xml; this commit adds them in.

Finally, `dellasm` may be set as one of the NTP servers, so its
definition in the hosts file is moved to happen before NTP is set.

[1]: https://support.microsoft.com/en-in/help/2385818/windows-time-service-doesn-t-start-automatically-on-a-workgroup-computer-that-s-running-windows-7-or-windows-server-2008-r2