ASM-9663 add network-online dependency to puppet.service
For RHEL/CentOS 7.3, there is a known bug that causes the puppet
service to run before the network is fully up. For most modules,
this won't be a real issue since puppet agent will retry again
later. However, this causes a problem for network module since
this module will try to rewrite the configurations and restart
the service, which results in misconfigured network service and
locked down service status for both network and puppet.

To fix this issue, the razor post_install script will now create
the custom puppet.service systemd script with network-online.target
for Requires and After sections in the service unit descriptor.

Puppet jira for this issue:
https://tickets.puppetlabs.com/browse/PUP-5402

UPDATE: Per comments on the PR, the post_install will use sed to
insert required segments and/or targets. Also, I corrected a
mispelled puppet service name in the puppetclean.service file.