LUD-77B enable razor broker to manage puppet installation
This commit removes logic, which installs and configures puppet,
from the OS specific task for redhat. Instead, the puppet related
processes are moved to broker. For this broker to work, the
appliance must have created the ludwig-puppet (this is done by
issuing razor command: "razor create-broker --name ludwig-puppet
--broker-type ludwig-puppet"). Along with this, there is a separate
change in asm-deployer that sets this broker to the newly created
razor policies during linux server deployment.