Point to razor OS repo URL as default yum repo
The CentOS ISO contains both the default RPM packages and GPG keys and
is available via razor's HTTP repo_url. This change sets that up as
the only enabled yum repository on a default CentOS install. This will
allow "yum install" commands to work out of the box even if the node
has no external internet access. This will make it easier to customize
the OS for different appliaction flows.