LUD-521 Only require PERC for RHEL for RAID case
The RHEL kickstart file always preferred to place the OS on a PERC
RAID volume. That should only be required if the `target_boot_device`
specifies `HD` (RAID).

This was breaking installs on VMs with PCI pass-through enabled and
PERC volumes -- the OS was being installed on one of the PERC volumes
rather than the first virtual disk.