LUD-547 Use dellasm hostname for razor-internal EL repo
Previously razor's `repo_url` was used to create a yum repository
pointing to the RHEL/CentOS repository files included on the
ISO. However that URL typically has an internal PXE IP which may not
be accessible by the node after the PXE network is removed.

Instead define and use an `/etc/hosts` file entry that points to the
internal PXE IP. It will be updated later on if needed by puppet
configuration when the PXE network is removed. The repository URL can
then use the `hosts` file entry instead of the IP directly.